+++ RealArtists.ShipHub.Api/Models/ApiEvent.cs
@@ -2,7 +2,7 @@
   using System;
 
   public class ApiEvent {
-    public int Identifier { get; set; }
+    public long Identifier { get; set; }
     public string CommitId { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
     public string Event { get; set; }
