+++ RealArtists.ShipHub.Common/DataModel/Types/RepositoryTableType.cs
@@ -1,7 +1,7 @@
 ﻿namespace RealArtists.ShipHub.Common.DataModel.Types {
   public class RepositoryTableType {
-    public int Id { get; set; }
-    public int AccountId { get; set; }
+    public long Id { get; set; }
+    public long AccountId { get; set; }
     public string Name { get; set; }
     public string FullName { get; set; }
     public bool Private { get; set; }
