+++ RealArtists.ShipHub.Database/User Defined Types/CommentTableType.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TYPE [dbo].[CommentTableType] AS TABLE (
-  [Id]           INT            NOT NULL,
+  [Id]           BIGINT         NOT NULL,
   [IssueNumber]  INT            NOT NULL,
-  [UserId]       INT            NOT NULL,
+  [UserId]       BIGINT         NOT NULL,
   [Body]         NVARCHAR(MAX)  NOT NULL,
   [CreatedAt]    DATETIMEOFFSET NOT NULL,
   [UpdatedAt]    DATETIMEOFFSET NOT NULL,
