+++ RealArtists.ShipHub.Database/User Defined Types/IssueEventTableType.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TYPE [dbo].[IssueEventTableType] AS TABLE (
-  [Id]            INT            NOT NULL PRIMARY KEY CLUSTERED,
+  [Id]            BIGINT         NOT NULL PRIMARY KEY CLUSTERED,
   [CreatedAt]     DATETIMEOFFSET NOT NULL,
   [ExtensionData] NVARCHAR(MAX)  NOT NULL
 )
