+++ RealArtists.ShipHub.Database/User Defined Types/IssueTableType.sql
@@ -1,16 +1,16 @@
 ﻿CREATE TYPE [dbo].[IssueTableType] AS TABLE (
-  [Id]           INT            NOT NULL PRIMARY KEY CLUSTERED,
-  [UserId]       INT            NOT NULL,
+  [Id]           BIGINT         NOT NULL PRIMARY KEY CLUSTERED,
+  [UserId]       BIGINT         NOT NULL,
   [Number]       INT            NOT NULL,
   [State]        NVARCHAR(6)    NOT NULL,
   [Title]        NVARCHAR(255)  NOT NULL,
   [Body]         NVARCHAR(MAX)  NULL,
-  [AssigneeId]   INT            NULL,
-  [MilestoneId]  INT            NULL,
+  [AssigneeId]   BIGINT         NULL,
+  [MilestoneId]  BIGINT         NULL,
   [Locked]       BIT            NOT NULL,
   [CreatedAt]    DATETIMEOFFSET NOT NULL,
   [UpdatedAt]    DATETIMEOFFSET NOT NULL,
   [ClosedAt]     DATETIMEOFFSET NULL,
-  [ClosedById]   INT            NULL,
+  [ClosedById]   BIGINT         NULL,
   [Reactions]    NVARCHAR(300)  NULL
 )
