+++ RealArtists.ShipHub.Database/User Defined Types/LabelTableType.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TYPE [dbo].[LabelTableType] AS TABLE (
-  [Id]    INT           NOT NULL,
+  [Id]    BIGINT        NOT NULL,
   [Color] NVARCHAR(6)   NOT NULL,
   [Name]  NVARCHAR(150) NOT NULL,
   PRIMARY KEY CLUSTERED ([Id], [Color], [Name])
