+++ RealArtists.ShipHub.Database/User Defined Types/RepositoryTableType.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TYPE [dbo].[RepositoryTableType] AS TABLE (
-  [Id]        INT            NOT NULL PRIMARY KEY CLUSTERED,
-  [AccountId] INT            NOT NULL,
+  [Id]        BIGINT         NOT NULL PRIMARY KEY CLUSTERED,
+  [AccountId] BIGINT         NOT NULL,
   [Private]   BIT            NOT NULL,
   [Name]      NVARCHAR(100)  NOT NULL,
   [FullName]  NVARCHAR(255)  NOT NULL
